[internal] Add new `dir:generator#generated` address syntax (#12891)
Part of https://docs.google.com/document/d/1HpJn2jTWf5sKob6zhe4SqHZ7KWBlX4a6OJOwnuw-IHo/edit#. As discussed there, we need a way to address generated targets so that we a) know which target generator to run to generate the target, and b) get namespacing for when target generators generate the same names.

We currently are using "file addresses" in the form `dir/f.ext:lib`, but that does not generalize to generating file-less targets like a `python_requirement_library`.

--

This adds the syntax: `dir:generator#generated`, e.g. `//:reqs#Django` or `src/py/project:lib#foo.py`.

Note that this goes left-to-right, from least specific to most specific. This fits with how depending on `dir:generator` results in depending on all the targets it generates, whereas `dir:generator#generated` is only the specific generated target.

We went with `#` for the delimiter, which aligns with URL syntax of `domain.com#fragment`, where `#` allows you to get a more specific part of the page. `!`, `?`, and `;` are special bash characters so need to be escaped. `:` would be confusing to recycle. `+` suggests addition when really this is making a big thing smaller. `=` reads weirdly. `@` is the only other option that makes sense.

--

CLI specs don't yet work with this syntax.

[ci skip-rust]
[ci skip-build-wheels]