[internal] Remove `go_external_module` and rename `_go_ext_module_package` to `_go_external_package` (#12924)
As explained in https://github.com/pantsbuild/pants/issues/12763, the atomic unit for Go 3rd-party dependencies is the `_go_external_package`, which is a subset of the `go_external_module` that users declare as top-level. This means that Pants always operates on `_go_external_package`, and the only use for `go_external_module` would be boilerplate reduction to generate the more atomic `_go_external_package`.

We will soon already have target generation for `_go_external_package` triggered by `go_module`, which reads the `go.mod`. So, the only use for `go_external_module` would be to declare deps that are not in your `go.mod`. But as discussed in #12763, we don't need to support that use case, at least for now.

[ci skip-rust]
[ci skip-build-wheels]