Build and install python_distribution deps for test, run, and repl (#12573)
If you (transitively) depend on a python_distribution, Pants will now build and
install it into the venv for test, run and repl.

We don't do anything special when packaging, under the assumption that the
python_distribution will itself be packaged and depended on as a published artifact.

This change allows you to seamlessly use existing/custom setup.py - notably to build Python
native extensions - in Pants. If a setup.py can build it, Pants can consume it.

Note that adding this support was very little code, thanks to the engine. Most of this
change is tests.

This change also tweaks the testprojects native extensions example, to clarify what
is going on.

Addresses #11942

[ci skip-rust]

[ci skip-build-wheels]