[internal] Reorganize Go code into `goals/`, `subsystems/`, and `util_rules/` (#12938)
This factoring has worked well for Python. `goals/` makes clear what is hooking up to goals like `package`, and it helps keep the surface area small for that code so we can put common functionality into `util_rules/`.

This is not the final factoring, but an incremental improvement.

Each commit is atomic and was automated with PyCharm's refactoring support.