[internal] Make more explicit when adding dependencies on siblings for generated file-level targets (#12906)
Due to https://github.com/pantsbuild/pants/pull/10582, we automatically were adding dependencies on all sibling files from a target generator (aka "build target") if that target did not have dependency inference, or its inference did not support inferring deps on siblings. This is important because users expect with languages like Python that every file in `python_library()` would automatically depend on every other file if it weren't for dependency inference.

But our heuristic was not fully accurate. For example, `files()` and `resources()` should never infer deps on siblings because there is no notion of a file depending on another file. If you need all the files, depend on the original `files()` target: only use file targets if you just need that single file. (That will be fixed in a followup.)

[ci skip-build-wheels]
[ci skip-rust]