Add Java dependency analysis types and launcher using javaparser library. (#12890)
This PR adds the core functionality for inferring from Java source:
* imports
* declared package
* top level types that could potentially be imported by other sources from the provided source

It uses the com.github.javaparser:javaparser-symbol-solver Java library to do the heavy lifting. We have a custom launcher for calling into this library, and that launcher source is compiled on-the-fly by Pants (the source itself is loaded using pkg_resources). This is a novel way of calling into Pants-provided JVM code that will probably act as a template for similar situations, like custom JUnit launchers.

[ci skip-rust]
[ci skip-build-wheels]