Remove `Enriched*Result` classes in favor of `EngineAwareReturnType.cacheable` (#12970)
The `Enriched*Result` classes were added in order to force re-rendering of high-level `lint`/`fmt`/`test`/`check` results, regardless of whether they had succeeded or failed. Although whether they should re-render even for successful results remains open to debate, #12877 allows for removing the "enriching" `@rule` and `Enriched*Result` wrappers by moving the specification of their cacheability directly onto the `(Lint|Fmt|Check|Test)Result` types.

When compared to the existing use of an "enriching" `@rule`, this change has the downside that the entire `@rule` body that produces a `*Result` will re-run. But because all of its inputs will be cached, this should be a trivial cost.