[internal] jvm/java: ensure JDK downloaded in one process (#12972)
## Motivation

As described in https://github.com/pantsbuild/pants/issues/12293, multiple Coursier invocations were downloading the JDK and triggering [a race condition in Coursier's locking](https://github.com/coursier/coursier/issues/1815) that caused flakiness in tests.

## Solution

This PR mitigates the issue by isolating JDK download to a single `Process`. The new `JdkSetup` type provides rules with the command to obtain the location of the JDK so they may query Coursier for JAVA_HOME. This has the benefit of still downloading in remote execution, but providing some guarantee that there will be a single download.