[internal] terraform: refactor parser script into its own file (#12985)
Move the parser script into its own file so that we can apply fmt/lint/check to the script's source. The Terraform plugin uses `pkgutil.get_data` to retrieve the module's source to store in the CAS and use as a PEX entry point.

The commented-out test for `resolve_pure_path` previously in `dependency_inference_test.py` can now be uncommented and placed in `hcl2_parser_test.py`.