[internal] go: enable plugin in repo and cleanup test project (#13018)
The Go plugin is far enough long that we should enable it in the repository. Also cleanup the BUILD files for the test project to match current plugin conventions. The test project's BUILD files are still ignored so that Pants developers do not need to install Go (due to target generation for go_module needing `go`).

[ci skip-rust]

[ci skip-build-wheels]