[internal] go: refactor compilation into separate rule (#13019)
Refactor invocation of the Go compiler into a separate rule with associated `CompileGoSourcesRequest` and `CompiledGoSources` types.

The motivation for this change is for test binary support where it will be useful to invoke the compiler once for non-test sources and again for test sources. With this change, the code won't duplicate compiler invocation logic.