[internal] go: add analyzer and rules for test sources (#13041)
Add a Go program to parse Go test sources and extract the names of tests and other metadata necessary for creating test binaries. It is based mainly on the [Bazel rules_go analyzer](https://github.com/bazelbuild/rules_go/blob/master/go/tools/builders/generate_test_main.go) and [`go` tool test analyzer](https://github.com/golang/go/blob/master/src/cmd/go/internal/load/test.go).

Note: It does not analyze ["testable examples"](https://go.dev/blog/examples) nor fuzz targets (which are unreleased and planned for Go 1.18+). Also, the `TestMain` detection needs to detect which package (internal or external test) the TestMain was in.

The main goal of this PR is to just get the basic framework in place for compiling and linking the analyzer binary. It is short on test cases which will come in future PRs as we iterate on the analyzer. (For example, there are some better analysis code in the Go distribution which will need to be incorporated.) But this PR is a good first start.

This PR in initial work for https://github.com/pantsbuild/pants/issues/11913.