[internal] Rename `go_module` target to `go_mod` (#13053)
A "module" is a generic concept in Go, and includes "external modules" like `github.com/google/uuid`. 

In contrast, our `go_module` target is specifically for first-party Go modules that are represented with a `go.mod` file. This change is meant to make more clear that connection, that the `go_mod` target should be used whenever you have a `go.mod` file. 

[ci skip-rust]
[ci skip-build-wheels]