[internal] Stop using `go.sum` when generating `_go_external_package` targets (#13052)
To create `_go_external_package` targets, all we need is the package import paths for an external module. We don't need any extra metadata like a package's dependencies and imports. We can find that by using `go list -find`, rather than `go list`.

By simplifying that, we no longer need to worry about having `go.sum` present when running `go list` to avoid `missing go.sum entry` errors. We can remove our hacky code that was merging the `go.sum` from disk (`go_mod` target) and any `go.sum` files that were created on-the-fly in the `DownloadedExternalModule` rule.

--

Related, we now pass the `go_sum_digest` to the call to `DownloadedExternalModule` so that it uses the `go.sum` from disk. This means we don't dynamically generate a `go.sum` on-the-fly, which is important for correctness: we should be using the lockfile the user intended for us to consume.

[ci skip-rust]
[ci skip-build-wheels]