[internal] Improve compatibility of nailgun with append only caches, and use them for Coursier (#13046)
Coursier is used to fetch JVMs and artifacts for Java support. On `main`, it uses its default cache directory, but as shown in #12293, this has concurrency issues when multiple clients are fetching JVMs at the same time.

The underlying issue is likely to be fixed by https://github.com/coursier/coursier/pull/2197, but in the meantime we can move to using `append_only_caches` for the `coursier` `--cache-dir` and `--jvm-dir`. Alone, this isn't sufficient to avoid concurrency issues (since fetches would simply collide in a new location): but it allows us to re-configure the cache location in `RuleRunner` tests using the `[pytest] execution_slot_var` to prevent collisions.

Along the way, support for mixing `append_only_caches` and `use_nailgun` needed fixing.

Re-enables running of JVM tests by default, and fixes #12293.