Add failing tests for Go external modules (#13065)
Our external module code does not consistently work, as shown by the test in `package_binary_integration_test.py`. Likewise, we are not properly loading the `go.sum` on disk when downloading dependencies, such as it having an invalid hash not erroring the `go mod download`.

This also adds a test for `ResolvedGoPackage` with external packages. For some reason, that test's package works even though the integration test fails in that same codepath.

Finally, this renames some test files to stop saying they're integration tests when they're really unit tests.

[ci skip-rust]
[ci skip-build-wheels]