[internal] Always use jars on the user classpath, and generalize transitive classpath building (#13061)
As described in #13036, Jars are faster to capture and materialize (fewer `inodes`), and have historically been significantly more performant for the JVM to load. We should have the output of JVM compilation be jars, always.

Additionally, generalize the construction of the transitive classpath needed to run an application via a `Addresses -> Classpath` `@rule`.

Fixes #13036.