[interal] Use `DownloadedExternalModules` when analyzing external Go packages (#13076)
Like https://github.com/pantsbuild/pants/pull/13070, this makes more progress on https://github.com/pantsbuild/pants/issues/12771.

We do not yet use `DownloadedExternalModules` for the sources digest used in `go_build_pkg.py`, only for the analysis of the package. That will be fixed in a followup.

To facilitate landing this, we remove the `go-pkg-debug` goal. It seems unlikely we'd want that for production, and you can simulate it by adding logging. It can be added back if we decide it's valuable once there is less churn.

[ci skip-rust]
[ci skip-build-wheels]