Refactor setup_py code to be more PEP 517 amenable in the future (#13096)
The code prior to this change was very enmeshed with the idea of generating an existing setup.py.
The ability to use an existing setup.py was tacked on, and the ability to use something
else (generated or existing) such as pyproject.toml/setup.cfg, let alone a different build backend
than setuptools, was nonexistent.

This change begins to tease apart the intermingled concepts.  

Notably, it deprecates the special-cased `setup_script` kwarg to the `setup_py()` BUILD file object. 
That was a hack with several drawbacks:

- It was out-of-place in an object that otherwise was just for specifying setup kwargs.
- It was mostly redundant in practice, since the only useful value it could take on was `"setup.py"`:
  Technically you could have it point to a `setup.py` file in some subdir beneath the target's dir,
  but that was really not useful in practice, since we cd into the dir containing setup.py when running setup,
  and it's highly unlikely anyone is using that mis-feature (especially since the ability to point to an existing 
  setup.py is new in 2.7.0).
- It meant you could only use an existing `setup.py`, and not some other existing build input such as `setup.cfg`.

Instead, we generate `setup.py` iff we don't find one in the expected location (i..e., in the target's dir). And in a 
followup we will add a boolean field to the target type itself to specify whether we should generate setup or not.

All this allows us to not pass setup_kwargs around, and use them only when we need them to generate a setup.py.
The one other non-test place we were using them was just to generate a unique directory name under dist for the
deprecated functionality of dumping the chroot. We just use the target path-safe spec instead, until the deprecated
code is deleted.

The result is simplified, clearer code, that provides a path to further teasing apart of setup.py-specific functionality
from general PEP-517 functionality, and allows for better encapsulation of logic inside single rules.

[ci skip-rust]

[ci skip-build-wheels]