[internal] Structure sharing for `CoarsenedTarget`s (#13087)
Profiles in example Java repositories showed (unsurprisingly) that recursively re-computing `CoarsenedTargets` (each of which did a transitive walk, and then discarded all but the `CoarsenedTarget`s for the input root) was a bottleneck.

Unlike `Targets`, `CoarsenedTargets` are by definition acyclic, and so can share structure. This change converts `CoarsenedTarget` to a transitive DAG, and `CoarsenedTargets` to containing only the roots of the walk.

Computing `CoarsenedTargets` no longer dominates compilation in example repos.