[internal] Stop setting `import_path="main"` for Go packages (#13113)
Given a project with `module = example.com/foo` in `go.mod`, and `hello.go` with `package main` in it, `go list -json` will report the `ImportPath` as `example.com/foo`, _not_ `main`. So it was wrong for us to be overriding the `import_path` to be `main`.

This is important for dependency inference to work properly when there are multiple `go.mod`s in the project. Dependency inference maps import paths to targets, and we'd have ambiguity if there were multiple `main` import paths.

This also fixes our calculation of `import_path` to not include a `/` suffix when the `package` is in the same directory as the `go.mod`, i.e. we should use `example.com/foo`, not `example.com/foo/`.

[ci skip-rust]
[ci skip-build-wheels]