Fix target generation for `terraform_modules` (#13131)
Generated targets "live" in the BUILD file of their target generator. That is, the address of `prod/terraform#subdir` has the `Address.spec_path` of `prod/terraform`, _not_ `prod/terraform/subdir` like one might expect. It's only a convention that `#subdir` happens to be a directory name, it could just as easily be called `#a_generated_tgt`.

Because the spec_path is the target generator's, we weren't properly relativizing the `sources` field.

This also renames the Target and Field names to better differentiate what is the Target API vs. internal abstractions.

[ci skip-rust]
[ci skip-build-wheels]