[internal] Fix `_go_internal_package` complaining about missing `.s` files (#13145)
We were explicitly setting the `sources` field to `[subdir/*.go, subdir/*.s]`. Because it's explicitly set, Pants errors/warns if any of the globs are not matched.

Instead, we set to the exact files used for that `_go_internal_package`. If the files change, the target will be regenerated.

This also partially implements https://github.com/pantsbuild/pants/issues/12953 to add a new `pkg_sources` field to allow the user to change what the triggering sources are, such as if they want to opt out of a certain directory. Note that the `go_mod` does not "own" these `Sources` field like normal, e.g. using a file arg on own of those files will not map back to the `go_mod`, only the `_go_internal_package`.

[ci skip-rust]