[internal] Rename Go package targets to `go_{first,third}_party_package` (#13165)
@tdyas pointed out that "internal" is a bad name because it makes Go developers think of `/internal/` in the import path, which is a common pattern.

Instead, some alpha Go users voted that they like best `first_party_package` and `third_party_package`.

This also makes those targets public by removing the `_`. We want them to show up in `./pants help` and the docs so that users understand what Pants is doing (if they so desire), and so that they know how to set the `overrides` field on `go_mod` when we add that. 

We use target validation to ensure that users never explicitly create these targets, as it's too unlikely things will break if so, like that they won't set some fields correctly or that their `go.mod` will be missing an entry.

[ci skip-rust]
[ci skip-build-wheels]