Actually fix Terraform target generation (#13181)
We weren't relativizing the `sources` field to the `terraform_modules` target, so it failed whenever that was not in the build root.

This also fixes the address when there are `.tf` files in the same directory as the target generator. It will be `src/tf#.`, going along with `src/tf#subdir`. Note that that contrast's with Go using `src/tf#./` and `src/tf#./subdir`, as Go developers will be familiar with the `./` syntax and that allows us to distinguish between first-party vs. third-party generated Go targets.

Finally, this switches two other target gen places to use `fast_relpath`, which simplifies the code and indeed should make it faster.

[ci skip-rust]
[ci skip-build-wheels]