[Internal] Invert `coursier_lockfile` dependency specifications (#13171)
This is the first step in providing swappable Coursier lockfiles: Instead of specifying input requirements directly to the coursier_lockfile target, jvm_artifacts are provided as dependencies for java_source(s) targets. The coursier-resolve goal now goes looking for targets that depend on the given lockfile, and includes any jvm_artifacts that are transitive dependencies of targets that depend on the lockfile.

This will allow individual source targets to specify multiple lockfiles as being compatible, without needing to manually ensure that each lockfile itself has the correct requirements.

Build-time artifacts are still materialised through the coursier lockfile, and so build-time performance should not be affected by these changes.

🚨🚨🚨 Note that this will break any in-progress JVM projects that you may be working on. 🚨🚨🚨

Relates to #12784.