`files` and `resources` targets now generate `file` and `resource` targets (#13235)
Now users can explicitly create a `file` and `resource` target. `files` and `resources` do that automatically.

### Impacts `./pants peek`

The `target_type` now says `file` and `resources`. There's no way around that, afaict.

### Impacts `./pants filter --target-type`

`--target-type=files` and `--target-type=resources` no longer matches the generated targets, which is breaking. I don't see how to fix this? For `python_source`, we could have it so that the deprecated name `python_library` matches both `python_sources` and `python_soruce`. But that isn't possible with `files` because there is no deprecated name, and it's wrong for `--target-type=files` to match `file`.

[ci skip-rust]
[ci skip-build-wheels]