[internal] Add `DirLiteralSpec` (#13219)
Allows us to represent when directories were specified, which is prework for https://docs.google.com/document/d/1WWQM-X6kHoSCKwItqf61NiKFWNSlpnTC5QNu3ul9RDk/edit#.

We don't actually do anything with this yet. For now, we always convert a `DirLiteralSpec` to `AddressLiteralSpec`. But it lays a groundwork for changing semantics.

This also adds a new way to specify the build root: `.`, e.g. `./pants lint .`. That also works if you do `./`.

[ci skip-rust]