Add `./pants update-build-files` to automate fixing some deprecation warnings (#13273)
For example:

```
❯ ./pants update-build-files
Updated src/python/pants/build_graph/BUILD:
  - Rename `python_library` to `python_sources`
Updated src/python/pants/core/goals/BUILD:
  - Rename `shell_library` to `shell_sources`
  - Rename `python_library` to `python_sources`
```

![Screen Shot 2021-10-16 at 9 13 21 AM](https://user-images.githubusercontent.com/14852634/137594612-ce5bf94a-8552-4bb8-9f42-1a51c24b6c35.png)

The functionality is generic, such that we (and plugin authors!) can add their own fixers. The API is simple: take in `lines: tuple[str, ..]` for the original BUILD file, and return `tuple[str, ...]`. Pants will handle piping each fixer onto the next, using the same technique as `./pants fmt`.

Includes a `--check` mode.

## Why `update-build-files`?

At first, we considered `mend` to go along with the sewing theme of `./pants tailor`. But a user pointed out that it can be harmful to use "cute" names like this, and @cczona also pointed out that `mend` has a negative connotation in sewing communities like that it was broken.

`update` and `fix` are too generic, as this is really only talking about changes to Pants's own BUILD files, not fixing your own code.

We considered `self-update`, but that sounds like this goal does more than it really does, like it changes the `pants_version` for you. Maybe we will add that functionality in the future! But for now, it's misleading. If we add it, we can always deprecate this goal in favor of the more powerful one.

[ci skip-rust]