`update-build-files` goal formats BUILD files with Black (#13279)
A frequent feature request is to autoformat BUILD files. 

I tried to do this with `./pants fmt` and `./pants lint` at first, but it proved extremely difficult to get working because it goes against our Target API / FieldSet code (no target "owns" the BUILD file) and results in weird Specs semantics.

Instead, `update-build-files` is a natural goal for this.

Regardless of if users have activated the Black backend already, they can set `[black].args`, `[black].config`, `[black].interpreter_constraints`, and `[python-setup].interpreter_search_paths` to influence how Black is installed and run.

--

We activate this autoformatting by default, partially based on feedback in Slack (10 votes for opt-out, 2 for opt-in).