Add `overrides` field to `shell_sources`, `shunit2_sources`, and `protobuf_sources` (#13298)
See https://github.com/pantsbuild/pants/pull/13270.

This does not yet add the field to `java_sources` and `junit_tests`, tracked by https://github.com/pantsbuild/pants/issues/13297.

I'm not sure if we should add this mechanism to `files` and `resources`? I'm having a hard time thinking of when you would want to override the metadata for a single file. But, we might want to do it for the sake of consistency?

This also does not hook up `python_requirements` and `poetry_requirements` because those use the old CAOF macro system, rather than the Target API.

[ci skip-rust]